requirements.txt
absl-py==0.7.1
astor==0.7.1
backcall==0.1.0
certifi==2019.3.9
cycler==0.10.0
Cython==0.29.7
decorator==4.4.0
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
imageio==2.5.0
imgaug==0.2.8
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
jupyter-client==5.2.4
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
labelme==3.11.2
Markdown==3.1
matplotlib==3.0.3
mock==2.0.0
networkx==2.2
numpy==1.16.2
opencv-contrib-python==4.0.0.21
opencv-python==4.0.1.24
parso==0.4.0
pbr==5.1.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
pygame==1.9.5
Pygments==2.3.1
pyparsing==2.3.1
PyQt5==5.12.1
PyQt5-sip==4.19.15
python-dateutil==2.8.0
PyWavelets==1.0.2
PyYAML==5.1
pyzmq==18.0.1
QtPy==1.7.0
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.2.1
Shapely==1.6.4.post2
six==1.12.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.2
termcolor==1.1.0
tornado==6.0.2
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.3
